pyproject.toml
[tool.poetry]
name = "postbox"
version = "0.1.0"
description = ""
authors = ["Eric Smaxwill <eric@smaxwill.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mangum = "^0.16.0"
fastapi = "^0.85.1"
pydantic = "^1.10.2"
aws-lambda-powertools = "^2.1.0"
email-validator = "^1.3.0"
cryptography = "^38.0.2"
httpx = "^0.23.0"
PyJWT = "^2.6.0"
devtools = "^0.9.0"
boto3 = "^1.25.4"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
uvicorn = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
